enhance(dependencies): Drop node 7 support
And declare minimum versions in package.json.

Node 7 was a dev release and is defunct now that 8 is LTS and 9 is the
current dev release. If we wanted to support earlier versions, we should
go all the way back to the prior LTS release series, 6, but I don't see
a reason to do that for new software.

Bump to npm 5 because that is the version that was released with node 8:
https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V8.md#8.0.0

This was prompted by the fact that current versions of yarn have dropped node 7
support.
https://yarnpkg.com/en/