style(scss): use function to apply z-index
This update uses a function to apply the z-index in each .scss file.
The function needs to iterate over a map of z-index values, so all of
the z-indexes end up in a single list and can be tracked more easily.

This allows us to manage all z-index values in a single location.