build(scripts): remove unneeded call to concurrently
There is an unneeded call to `concurrently` in the `npm run dev` script.
This causes things that should be output in colour (eg debug messages
when running with DEBUG=*) to not be output in colour which makes
debugging things harder than it should be.

This was in place because previously lnd was being started up externally
in the `npm run dev` script at the same time as starting up the app. But
we now use spawn from within the app to start up lnd.

The call to `concurrently` is no longer needed.

See https://github.com/LN-Zap/zap-desktop/commit/d8c5623e0131d89e1fb889eb5356fd1752bcbf12