build(lint): lint/format in git pre-commit hook
Currently we use eslint and prettier to lint and format code. However,
running the linter is a manual step that is easy to forget.

To improve this we run our lint scripts in a git pre-commit hook. This
functionality is optional and can be enabled by adding a .opt-in file
with a line `pre-commit`.