refactor: modularize and simplify app startup code
`main.dev.js` was hard to follow and understand the code flow. It housed
code to handle quite a few distinct things. This is a fairly substantial
refactor of the application startup code in which we:

A new `Neutrino` class has been created with 2 public methods. `start`
and `stop`. `start` will launch a new `lnd` process whilst `stop` will
stop it. This class extends the `EventEmitter` class and emits the
following events based on activity detected from the lnd log output.

 - `grpc-proxy-started` - gRPC started
 - `wallet-opened` Wallet opened
 - `fully-synced` - lnd is all caught up to the blockchain
 - `got-block-height` - got updated block height

A new `ZapController` class  has been created which houses all of the
logic for intraprocess communication between the main and renderer
processes.

Previously we had several `setInterval` loops that were checking every
second to see if the application status has changed and trigging the
appropriate action if so. This was pretty hard to follow has been
replaced here with more extensive use of promises. This enables us to
act instantly to relevant changes rather than waiting up to 1 second
for the next interval to fire.

Now, the only stuff that lives in `main.dev.js` now is the top level
`app` listeners, which calls out the other parts mentioned above to
bootstrap the application.