ci: enable electron security checks
Unset the `ELECTRON_DISABLE_SECURITY_WARNINGS` environment variable
which was being used to disable the security checks when testing the app
on ci. This is no longer needed now that we have a more strict Content
Security Policy when the app runs in production mode.

See https://github.com/LN-Zap/zap-desktop/pull/456