fix(wallet): ensure clean state after refresh
The `onOnboarding` state transition handler only fires if you are
entering the `onboarding` state from another state. If you are already
in the `onboarding` state it does not fire when the `startOnboarding`
event is initiated. By contrast, `onStartOnboarding` always fires when
`startOnboarding` is run.

Ensure that IPC listeners are only registered once when transitioning to
the `onboarding` state from another state.