fix(ui): ensure content size same across platforms
We set `titleBarStyle: hidden` which causes the title bar to be hidden
on the mac so that we can better integrate it with our designs. On
windows this has no effect and the titlebar is still rendered. This
causes the actual display area to be different on mac vs windows.

Set the `useContentSize` to ensure that the rendered content is the
same size on all platforms.

See https://github.com/electron/electron/blob/master/docs/api/browser-window.md#new-browserwindowoptions

Fix #666