fix(grpc): close connection after failed connect
The gRPC Client waitForReady call waits for a specified time frame and
either resolves or rejects depending on wether or not a successful
connection was established within the given time frame. If the
connection attempt failed gRPC will still continue to attempt to
establish the connection endlessly.

In this commit we explicitly close the client connection after the
connect deadline has been exceeded to prevent gRPC from trying to
connect.