test(coverage): disable text coverage reporter
The txt coverage reporter generates a lot of output that makes it hard
to see the actual test results without scrolling up a lot first.
We already have the html reporter setup and an npm script to enable
easy viewing of the data (`yarn coverage`) so the text reporter isn't
adding anything useful.