test(jest): run jest with --forceExit flag
Jest is sometimes unable to detect when tests have completed which can
result in tests hanging indefinitely, which breaks our CI builds on
Appveyor.

This is a know issue in jest and they added a `forceExit` in order to
help users facing this issue.

See https://github.com/facebook/jest/issues/1456