fix(wallet): hide expired toggle for empty wallet
Hide the "show / hide expired" toggle for new users with no previous
activity.

For new users with an empty wallet it doesn't make any sense that in
middle of the activity list there is a toggle button to show/hide
expired requests, since we know the user has no requests.