fix(boot): issue with fetching user settings
When we start up the app we attempt to fetch the users store settings
from IndexedDb by loading up a hidden browser window and execute
code within it to connect to and fetch settings from the database.

This fix ensures that we properly close out this hidden window once we
have fetched the user settings. This resolves an issue that prevented
the production build from starting up fully.