feat(grpc): use @grpc/grpc-js grpc implementation
Switch to the @grpc/grpc-js package, a pure javascript implementation,
to simplify greatly the installation story for the end users.

The grpc package contains native code, and causes grief and difficulty
of usage for the end user. Since electron 3, it became possible to use
the pure javascript version of grpc, since the necessary http2 module
became available there.

Fix #951