fix(channels): refresh channel details on update
Fetch an updated channel list if we receive a channel update to one of
our active channels.

Previously, we would only update if the advertising_node or
connecting_node was our own node. This fix expands the criteria so that
we also trigger the channels list re-fetch if there is a match against
one of our channels' remote_pubkey.

Fix #724