build(electron): set autoupdate channel to "beta"
Set the update channel to `beta` so that users will be only be upgraded
to "beta" or "latest" versions of the app automatically.

This frees up the alpha channel to be used by developers or users that
specifically enable it. Allow overriding the update channel by starting
the app with the `AUTOUPDATE_CHANNEL` environment variable set.