fix(app): reset onboarding after app refresh
Only trigger the onboarding process to start once the app is ready. This
means waiting until we have loaded all the available wallet data and
the onboarding process has been triggered from the main process.

This ensures that we do not attempt to start the onboarding process
until the main process has been fully reset (disconnect / stop lnd) and
prevents us from trying to start lnd when it is already started after
a window refresh.