fix(app): ensure app quits fully on osx
On mac, when the main window is closed we hide the app instead of
triggering the app quit process. When the user actually quits the app
(from the menu bar or pressing ctrl+q) we set a `forceClose` property
on the window object.

Honour the `forceClose` property in the `window-all-closed` event to
ensure the app properly quits on mac as implementing this hook prevents
the electron default behaviour of quitting the app.