fix(onboarding): handle host/cert errors on boot
When the app starts it tries to automatically log into the last active
wallet. If there is a problem with the cert or macaroon files, the user
is left hanging on the loading screen.

Consolidate errors from startLnd into a single object, and ensure that
we handle all lnd start errors.

Fix #1215