ci(node): do not build and test on node LTS
Our current CI process is relatively time consuming, largely due to the
fact that we currently build and test the app on node LTS in addition to
the current node version.

Node LTS primarily exists to server enterprises with an inability to
move fast and keep technology current. We don't fit into that category
and should waste time building and testing the app on an old node LTS
version.