fix(wallet): correctly identify active network
Due to recent changes in lnd network information (chain/network) can
be located in different parts of the lnd getInfo api call response. We
have already taken some measures to detect and modify this information
so that it is consistent but there were some places in our code where
we were still looking for the info in the old location and as a result
incorrectly identifying the network that a node is running on.

Fix #1510