fix(app): retain wallets & settings data on logout
Wallets and settings data is only only fetched once, when the Root
component first mounts. This data was getting deleted when logging out
of a wallet as part of a wider state reset. Exclude the settings and
wallets reducers from the state reset so that the data is retained
throughout the app lifecycle.

Fix #1723