fix(lnd): don't reset lnd state on hot-reload
When the app hot reloads (in development mode), the lnd state is
currently being reset and the user is sent back to the launchpad. This
causes a broken app state since lnd is actually still running but the
app thinks that it isn't.

Do not reset the lnd state when DialogLndCrashed unmounts (this only
happens during hot reload).