fix(db): resolve Dexie error on app startup
Initialise Dexie in the app preload script. This ensures that it is
initialised as early as possible and resolves an error that was caused
by trying to include the library from the app index file. It's also
cleaner this way as it consolidates our app initialisation code into a
single place.