feat(onboarding): ability to import encrypted seed
If the import fails with an error, check the error message to determine
the reason for the failure. If the reason was that an invalid passphrase
was provided, present the user with a password field where they can
enter the cipher seed passphrase.

This does not change the standard import flow. It adds a new step to the
end of the process only if we detect that the import failed due to a
missing passphrase.

Fix #2071