fix(ui): separate version/commit on profile pane
When connected to a node that is running a custom lnd build the lnd
version string as reported by `getInfo` is quite long since it contains
the version string and commit identifier.

This causes a display glitch on the node profile page when viewed at
the default window size.

To address this, split the version string out from the commit
identifier and display over two lines with the commit identifier in a
smaller lighter font.