fix(ui): ensure active wallet is set early
In order to effectively render the wallet we must know which wallet is
being rendered (which wallet is the active wallet). This chance ensures
that we set the current active wallet before registering that the lnd
gRPC Lightning interface has been successfully connected to, since this
is the trigger that use for redirecting between the onboarding process
and the wallet screen.